Ultimo comit
Se trato de configurar para trabajarlo por procedimiento funciones etc
mas no se logro del todo alcanzar el objetivo en cuanto a las cantidades

+++ src/main/java/com/concentrados/Danny/service/ItemServiceImpl.java
@@ -97,7 +97,7 @@ public class ItemServiceImpl implements ItemService {
         List<Item> lista = getItems();
         BigDecimal total = BigDecimal.ZERO;
         for (Item i : lista) {
-            total = total.add(i.getSubTotal());
+            total = total.add(BigDecimal.valueOf(i.getSubTotal()));
         }
         return total.doubleValue();
     }
